Moved shared code for downloading and reading the Pilot 2 data into the p2_common.py file.  Moved the p2b2_keras.py file back to the p2b2.py file.
