+++ Pilot2/P2B1/p2b1_baseline_keras2.py
@@ -3,7 +3,6 @@ import numpy as np
 import scipy as sp
 import pickle
 import sys,os
-import glob
 import argparse
 file_path = os.path.dirname(os.path.realpath(__file__))
 lib_path = os.path.abspath(os.path.join(file_path, '..', 'common'))
@@ -15,7 +14,7 @@ from keras import backend as K
 
 from data_utils import get_file
 
-import p2b1 as p2b1 # BVE go back and look at this
+import p2b1 as p2b1
 import p2_common as p2c
 import p2_common_keras as p2ck
 
@@ -28,7 +27,7 @@ def str2bool(v):
 def get_p2b1_parser():
         parser = argparse.ArgumentParser(prog='p2b1_baseline',
             formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-            description='Train Molecular Frame Auto-encoder - Pilot 2 Benchmark 1')   
+            description='Train Molecular Frame Autoencoder - Pilot 2 Benchmark 1')   
 
         return p2b1.common_parser(parser)
 
@@ -66,8 +65,6 @@ def run(GP):
 
     import keras_model_utils as KEU
     reload(KEU)
-    import pilot2_datasets as p2
-    reload(p2)
     reload(p2ck)
     maps=hf.autoencoder_preprocess()
 	
@@ -84,14 +81,8 @@ def run(GP):
     kerasDefaults = p2c.keras_default_config()
 	
 ##### Read Data ########
-    print ('Reading Data...')
-    data_set=p2.data_sets[GP['set_sel']][0]
-    data_hash=p2.data_sets[GP['set_sel']][1]
-    print ('Reading Data Files... %s->%s' % (GP['set_sel'], data_set))
-    data_file = get_file(data_set, origin='http://ftp.mcs.anl.gov/pub/candle/public/benchmarks/Pilot2/'+data_set+'.tar.gz', untar=True, md5_hash=data_hash)
-    data_dir = os.path.join(os.path.dirname(data_file), data_set)
-    data_files=glob.glob('%s/*.npy'%data_dir) 
-    
+    data_files=p2c.get_list_of_data_files(GP)
+        
     ## Define datagenerator
     datagen=hf.ImageNoiseDataGenerator(corruption_level=GP['noise_factor'])
